Introduce a `UserEmbedder` class
