+++ src/poprox_recommender/embedders/__init__.py
@@ -1,3 +1,4 @@
 from poprox_recommender.embedders.article import ArticleEmbedder
+from poprox_recommender.embedders.user import UserEmbedder
 
-__all__ = ["ArticleEmbedder"]
+__all__ = ["ArticleEmbedder", "UserEmbedder"]
